+++ README.md
@@ -20,7 +20,7 @@ The aim of this project is to build a feature-rich, easy to use, and highly perf
 
 While our distributed STARK prover implementation is still under development, this project contains fully-functional, multi-threaded, STARK prover and verifier code with the following nice properties:
 
-**A simple interface.** This library provides a relatively simple interface for describing general computations. See [usage](#Usage) for a quick tutorial, [common crate](common) for the description of the interface, and [examples crate](examples) for a few real-world examples.
+**A simple interface.** This library provides a relatively simple interface for describing general computations. See [usage](#Usage) for a quick tutorial, [air crate](air) for the description of the interface, and [examples crate](examples) for a few real-world examples.
 
 **Multi-threaded proof generation.** When compiled with `concurrent` feature enabled, the proof generation process will run in multiple threads. The library also supports concurrent construction of execution trace tables. The [performance](#Performance) section showcases the benefits of multi-threading.
 
@@ -46,8 +46,8 @@ The project is organized into several crates like so:
 | [examples](examples) | Contains examples of generating/verifying proofs for several toy and real-world computations. |
 | [prover](prover)     | Contains implementations of STARK provers. You can use these to generate computational integrity proofs. |
 | [verifier](verifier) | Contains an implementation of a STARK verifier which can verify proofs generated by a prover. |
+| [air](air)           | Contains components needed to describe arbitrary computations in a STARK-specific format. |
 | [fri](fri)           | Contains implementation of a FRI prover and verifier. These are used internally by the STARK prover and verifier. |
-| [common](common)     | Contains modules with components shared between the prover and the verifier. |
 | [math](math)         | Contains modules with math operations needed in STARK proof generation/verification. These include: finite field arithmetic, polynomial arithmetic, and FFTs. |
 | [crypto](crypto)     | Contains modules with cryptographic operations needed in STARK proof generation/verification. Specifically: hash functions and Merkle trees. |
 | [utils](utils)       | Contains a few utility functions used throughout the library. |
@@ -115,13 +115,13 @@ pub fn build_do_work_trace(start: BaseElement, n: usize) -> ExecutionTrace<BaseE
 }
 ```
 
-Next, we need to define the AIR for our computation. We do this by implementing the `Air` trait. At the high level, the code below does three things:
+Next, we need to define *algebraic intermediate representation* (AIR) for our computation. This process is usually called *arithmetization*. We do this by implementing the `Air` trait. At the high level, the code below does three things:
 
 1. Defines what the public inputs for our computation should look like. These inputs are called "public" because they must be known to both, the prover and the verifier.
-2. Defines a transition function with a single transition constraint. This transition constraint must evaluate to zero for all valid state transitions, and to non-zero for any invalid state transition. The degree of this constraint is 3 (see more about constraint degrees [here](common/#Constraint-degrees)).
-3. Define two assertions against an execution trace of our computation. These assertions tie a specific set of public inputs to a specific execution trace (see more about assertions [here](common/#Trace-assertions)).
+2. Defines a transition function with a single transition constraint. This transition constraint must evaluate to zero for all valid state transitions, and to non-zero for any invalid state transition. The degree of this constraint is 3 (see more about constraint degrees [here](air/#Constraint-degrees)).
+3. Define two assertions against an execution trace of our computation. These assertions tie a specific set of public inputs to a specific execution trace (see more about assertions [here](air/#Trace-assertions)).
 
-For more information about the `Air` trait see [common crate](common#Air-trait), and here is the actual code:
+For more information about arithmetization see [air crate](air#Arithmetization), and here is the actual code:
 
 ```Rust
 use math::field::{f128::BaseElement, FieldElement};
