+++ prover/src/lib.rs
@@ -8,7 +8,7 @@ pub use monolith::{prove, ExecutionTrace, ExecutionTraceFragment};
 
 mod channel;
 
-pub use common::{
+pub use air::{
     proof::StarkProof, Air, AirContext, Assertion, EvaluationFrame, FieldExtension, HashFunction,
     ProofOptions, TraceInfo, TransitionConstraintDegree, TransitionConstraintGroup,
 };
