+++ prover/src/monolith/constraints/commitment.rs
@@ -3,7 +3,7 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-use common::proof::Queries;
+use air::proof::Queries;
 use crypto::{ElementHasher, MerkleTree};
 use math::FieldElement;
 use utils::{batch_iter_mut, uninit_vector};
