+++ prover/src/monolith/domain.rs
@@ -3,7 +3,7 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-use common::Air;
+use air::Air;
 use math::{fft, log2, StarkField};
 
 // TYPES AND INTERFACES
