+++ verifier/Cargo.toml
@@ -18,4 +18,4 @@ utils = { path = "../utils", package = "winter-utils" }
 math = { path = "../math", package = "winter-math" }
 crypto = { path = "../crypto", package = "winter-crypto" }
 fri = { path = "../fri", package = "winter-fri" }
-common = { path = "../common", package = "winter-common" }
+air = { path = "../air", package = "winter-air" }
