+++ verifier/src/composer.rs
@@ -3,7 +3,7 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-use common::{Air, DeepCompositionCoefficients, EvaluationFrame, FieldExtension};
+use air::{Air, DeepCompositionCoefficients, EvaluationFrame, FieldExtension};
 use math::FieldElement;
 
 // DEEP COMPOSER
