+++ verifier/src/evaluator.rs
@@ -3,7 +3,7 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-use common::{Air, ConstraintCompositionCoefficients, EvaluationFrame};
+use air::{Air, ConstraintCompositionCoefficients, EvaluationFrame};
 use math::{polynom, FieldElement};
 
 // CONSTRAINT EVALUATION
